pyproject.toml
[project]
name = "disktrend"
version = "1.0.0"
description = "Linux storage analyzer with web dashboard"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "apscheduler>=3.10.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "aiosqlite>=0.20.0",
    "psutil>=6.0.0",
]

[project.scripts]
disktrend = "disktrend.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/disktrend"]
